ticket #330 Content windows: java.lang.NumberFormatException: For input string "0.1"

+++ src/com/sqleo/environment/ctrl/content/TaskUpdate.java
@@ -20,6 +20,7 @@
 
 package com.sqleo.environment.ctrl.content;
 
+import java.math.BigDecimal;
 import java.sql.PreparedStatement;
 import java.sql.Types;
 import java.sql.Savepoint;
@@ -227,7 +228,9 @@ public class TaskUpdate implements Runnable
 				pstmt.setInt(i+1, new Integer(param[0].toString()).intValue());
 				break;
 			case Types.NUMERIC:
-				pstmt.setLong(i+1, new Long(param[0].toString()).longValue());
+//				pstmt.setLong(i+1, new Long(param[0].toString()).longValue());
+				pstmt.setBigDecimal(i+1, new BigDecimal(param[0].toString()));
+//				pstmt.setDouble(i+1, new Double(param[0].toString()).doubleValue());
 				break;
 			default:
 				pstmt.setObject(i+1,param[0],((Integer)param[1]).intValue());
